Crate pbjson_types

pbjson-types provides the google.protobuf types, commonly known as well-known-types, with serde::Serialize and serde::Deserialize implementations that are compliant with the protobuf JSON mapping

Note: Coverage of all types is currently incomplete, some may have non-compliant implementations

Modules

descriptor_proto

Nested message and enum types in DescriptorProto.

enum_descriptor_proto

Nested message and enum types in EnumDescriptorProto.

extension_range_options

Nested message and enum types in ExtensionRangeOptions.

feature_set

Nested message and enum types in FeatureSet.

feature_set_defaults

Nested message and enum types in FeatureSetDefaults.

field

Nested message and enum types in Field.

field_descriptor_proto

Nested message and enum types in FieldDescriptorProto.

field_options

Nested message and enum types in FieldOptions.

file_options

Nested message and enum types in FileOptions.

generated_code_info

Nested message and enum types in GeneratedCodeInfo.

method_options

Nested message and enum types in MethodOptions.

source_code_info

Nested message and enum types in SourceCodeInfo.

uninterpreted_option

Nested message and enum types in UninterpretedOption.

value

Structs

Any

Api

BoolValue

BytesValue

DescriptorProto

DoubleValue

Duration

Empty

Enum

EnumDescriptorProto

EnumOptions

EnumValue

EnumValueDescriptorProto

EnumValueOptions

ExtensionRangeOptions

FeatureSet

FeatureSetDefaults

Field

FieldDescriptorProto

FieldMask

FieldOptions

FileDescriptorProto

FileDescriptorSet

FileOptions

FloatValue

GeneratedCodeInfo

Int32Value

Int64Value

ListValue

MessageOptions

Method

MethodDescriptorProto

MethodOptions

Mixin

OneofDescriptorProto

OneofOptions

Option

ServiceDescriptorProto

ServiceOptions

SourceCodeInfo

SourceContext

StringValue

Struct

Timestamp

Type

UInt32Value

UInt64Value

UninterpretedOption

Value

Enums

Edition

NullValue

Syntax